Add implementation of example for visitor patern
Part I: Implement Three Different Visitors
Each class has three methods you must write implementations for:

1. getResult(): Return an integer denoting the result, which is different for each class:

The SumInLeavesVisitor implementation must return the sum of the values in the tree's leaves only.
The ProductRedNodesVisitor implementation must return the product of values stored in all red nodes, including leaves,
 computed modulo 10^9 + 7. Note that the product of zero values is equal to 1.
The FancyVisitor implementation must return the absolute difference between the sum of values stored in the tree's non-leaf nodes at even depth and the sum of values stored in the tree's green leaf nodes. Recall that zero is an even number.
2. visitNode(TreeNode node): Implement the logic responsible for visiting the tree's non-leaf nodes such that the getResult method returns the correct result for the implementing class' visitor.

3. visitLeaf(TreeLeaf leaf): Implement the logic responsible for visiting the tree's leaf nodes such that the getResult method returns the correct result for the implementing class' visitor.

Part II: Read and Build the Tree
Read the n-node tree, where each node is numbered from 1 to n. The tree is given as a list of node values (x1, x2, ..., xn), a list of node colors (c1, c2, ..., cn), and a list of edges. Construct this tree as an instance of the Tree class. The tree is always rooted at node number 1.
----------------------
Sample Input

5
4 7 2 5 12
0 1 0 0 1
1 2
1 3
3 4
3 5
----------------------
Sample Output

24
40
15